- The launcher application for starting pygame apps -- TODO: Symbian does not let one share the open (font) files thus there can be only 1 active font at a time. This means that fonts cannot be cached. The images must be cached instead. Menu update is currently noticeably slow because each time a font is used, it is loaded from file.
- liquid_s60.py, the liguid effect

+++ symbian/BuildPygameRelease.py
@@ -55,13 +55,14 @@ def build():
     args = { "applications" : "",
              "capabilities" : CAPABILITIES,
              "builtin"      : "sysinfo,socket",
+             #"pyds"         : "gles",
              "basename"     : "pygame_python",            
              "uidbase"      : hex(UID_BASE).replace("L",""),             
              "sisappname"   : '"Python for Pygame"',
              # Convert to int or may be converted to octal due to zero at beginning
              'sisversion'   : '"(1,%d,%d%s)"' % ( int(version[2:4]), int( version[4:6]),version[6:]),
              'pythonsis'    : sisname,
-             'libpath'      : "data/pygame/libs",
+             'libpath'      : "data/pygame/libs",             
              }
     
     # Add certificate stuff
@@ -74,16 +75,19 @@ def build():
     sisname   = ""  
     if config.build_python:
         curdir = os.getcwd()
-        os.chdir(config.pys60_ce_src)                
-        dobuild(args)    
+        os.chdir(config.pys60_ce_src)   
+        a = args.copy()
+        a["gcce_options"] = '"-O2 -fno-unit-at-a-time"'             
+        dobuild(a)    
         os.chdir(curdir)
     
         sisname = "python_for_pygame_%s_signed.sisx" % version
         pys60_sis = os.path.join( config.pys60_ce_src, sisname )
                 
         # Copy the sis to current directory
-        import shutil
-        shutil.copyfile(pys60_sis, sisname)
+        if os.path.exists(pys60_sis):
+            import shutil
+            shutil.copyfile(pys60_sis, sisname)
         
         args['pythondll'] =  args['basename']
